Support up to 500 character file names
Cherry-pick from master (0bae3bb):
- Increase file name buffer from 128 to 500 characters
- Affects SleepActivity, FileSelectionActivity, CrossPointWebServer

Fixes long filename display issues.

+++ src/activities/reader/FileSelectionActivity.h
@@ -14,7 +14,7 @@ class FileSelectionActivity final : public Activity {
   SemaphoreHandle_t renderingMutex = nullptr;
   std::string basepath = "/";
   std::vector<std::string> files;
-  int selectorIndex = 0;
+  size_t selectorIndex = 0;
   bool updateRequired = false;
   const std::function<void(const std::string&)> onSelect;
   const std::function<void()> onGoHome;
